New and improved component system

+++ Relic/source/Relic/Core/Application.h
@@ -35,7 +35,7 @@ namespace Relic
         void Run();
         void Close();
 
-        void Constrain(std::shared_ptr<Entity> entity, uint32_t x, uint32_t y);
+        void Constrain(std::shared_ptr<Entity> entity, float x, float y);
         void Draw(const sf::Drawable& drawable);
 
         inline const std::string& GetTitle() const { return m_window->GetTitle(); }
